Use Unicorn as server for everything
Before this we had a very strange setup using passenger with
a nginx being the middleware for an apache server.

We stop this madness; we now use unicorn for the entire
application, including the static files (performance is not
really an issue).

We remove the thin and the unicorn gems. Also the commented
forman gem oO